// projects/sx_lib_front/src/lib/enums/snackbar-position.enum.ts
export enum SnackbarPosition {
    TOP_LEFT = 'top-left',
    TOP_CENTER = 'top-center',
    TOP_RIGHT = 'top-right',
    BOTTOM_LEFT = 'bottom-left',
    BOTTOM_CENTER = 'bottom-center',
    BOTTOM_RIGHT = 'bottom-right',
    CENTER_LEFT = 'center-left',
    CENTER = 'center',
    CENTER_RIGHT = 'center-right'
} 